updated
This Cloudformation template will create a Lambda Function for the Custom Config Rule, which checks security groups do not have an inbound rule with a port range of "All".

You can use this cloud formation template to create a config rule in AWS config. This can enable compliance auditing, security analysis, and change management for the Security Group.